back/requirements.txt
click==7.1.2
ecdsa==0.14.1
Flask==1.1.2
Flask-Cors==3.0.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pyasn1==0.4.8
python-dotenv==0.14.0
python-jose==3.2.0
redis==3.5.3
rsa==4.6
six==1.15.0
Werkzeug==1.0.1
